NCL Theme Gradient colour Scale

Generates a two colour gradient (low-high)

Usage

scale_colour_ncl(
  palette = c("ics", "icb", "icb_darkblue", "icb_brightblue", "icb_aquablue",
    "icb_lightblue", "icb_nhsblue", "icb_darkgrey", "icb_midgrey", "ics_purple",
    "ics_pink", "ics_limegreen", "ics_brightgreen", "ics_teal"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_color_ncl(
  palette = c("ics", "icb", "icb_darkblue", "icb_brightblue", "icb_aquablue",
    "icb_lightblue", "icb_nhsblue", "icb_darkgrey", "icb_midgrey", "ics_purple",
    "ics_pink", "ics_limegreen", "ics_brightgreen", "ics_teal"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_fill_ncl(
  palette = c("ics", "icb", "icb_darkblue", "icb_brightblue", "icb_aquablue",
    "icb_lightblue", "icb_nhsblue", "icb_darkgrey", "icb_midgrey", "ics_purple",
    "ics_pink", "ics_limegreen", "ics_brightgreen", "ics_teal"),
  discrete = TRUE,
  reverse = FALSE,
  ...
)

Arguments

palette

The name of palette to use

discrete

Optional: boolean to indicate that this scale is a discrete scale. Defaults to TRUE

reverse

Optional: boolean to reverse the direction of the scale. Defaults to FALSE

...

additional arguments passed to the ggplot functions

Value

a scale object to be used with a ggplot object