Creates the NCL ICS Theme for ggplot

Creates a base theme for ggplot objects. By default this uses a font that requires installing, and this is handled in the package load and `ncl_setup` function package to be installed.

Usage

theme_nclics(base_family = "Mulish")

Arguments

base_family

The font to use in this theme. This defaults to 'Mulish', but you can change it to any font you have installed.

See also

Examples

if (FALSE) { # \dontrun{
library(ggplot2)
library(NCLRtemplates)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point() +
  theme_nclics()
} # }