Convert to an NCL Word document

Format for converting from R Markdown to a NCL ICB branded MS Word document.

Usage

ncl_document(
  use_numbered_headings = FALSE,
  toc = FALSE,
  toc_depth = 1,
  fig_caption = TRUE,
  df_print = "kable",
  highlight = "default",
  keep_md = FALSE,
  md_extensions = NULL,
  pandoc_args = NULL
)

Arguments

use_numbered_headings

should we use numbered headings? Defaults to false

toc

TRUE to include a table of contents in the output

toc_depth

Depth of headers to include in table of contents

fig_caption

TRUE to render figures with captions

df_print

Method to be used for printing data frames. Valid values include "default", "kable", "tibble", and "paged". The "default" method uses a corresponding S3 method of print, typically print.data.frame. The "kable" method uses the knitr::kable function. The "tibble" method uses the tibble package to print a summary of the data frame. The "paged" method creates a paginated HTML table (note that this method is only valid for formats that produce HTML). In addition to the named methods you can also pass an arbitrary function to be used for printing data frames. You can disable the df_print behavior entirely by setting the option rmarkdown.df_print to FALSE. See Data frame printing section in bookdown book for examples.

highlight

Syntax highlighting style passed to Pandoc.

Supported built-in styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", "haddock", and "breezedark".

Two custom styles are also included, "arrow", an accessible color scheme, and "rstudio", which mimics the default IDE theme. Alternatively, supply a path to a .theme file to use a custom Pandoc style. Note that custom theme requires Pandoc 2.0+.

Pass NULL to prevent syntax highlighting.

keep_md

Keep the markdown file generated by knitting.

md_extensions

Markdown extensions to be added or removed from the default definition of R Markdown. See the rmarkdown_format for additional details.

pandoc_args

Additional command line options to pass to pandoc

Value

R Markdown output format to pass to render

Details

See the online documentation for additional details on using the word_document format.

Examples

if (FALSE) {
library(rmarkdown)

# simple invocation
render("input.Rmd", ncl_document())
}